Refactor test file
