+++ python_modules/dagster/dagster_tests/asset_defs_tests/test_antlr_asset_selection.py
@@ -2,6 +2,9 @@ import pytest
 from dagster._core.definitions.antlr_asset_selection.antlr_asset_selection import (
     AntlrAssetSelectionParser,
 )
+from dagster._core.definitions.asset_selection import AssetSelection, CodeLocationAssetSelection
+from dagster._core.definitions.decorators.asset_decorator import asset
+from dagster._core.storage.tags import KIND_PREFIX
 
 
 @pytest.mark.parametrize(
@@ -63,3 +66,71 @@ from dagster._core.definitions.antlr_asset_selection.antlr_asset_selection impor
 def test_antlr_tree(selection_str, expected_tree_str):
     asset_selection = AntlrAssetSelectionParser(selection_str)
     assert asset_selection.tree_str == expected_tree_str
+
+
+@pytest.mark.parametrize(
+    "selection_str",
+    ["not", "a b", "a and and", "a and", "sinks", "owner", "tag:foo=", "owner:owner@owner.com"],
+)
+def test_antlr_tree_invalid(selection_str):
+    with pytest.raises(Exception):
+        AntlrAssetSelectionParser(selection_str)
+
+
+@pytest.mark.parametrize(
+    "selection_str, expected_assets",
+    [
+        ('"a"', AssetSelection.assets("a")),
+        ("not a", AssetSelection.all() - AssetSelection.assets("a")),
+        ("a and b", AssetSelection.assets("a") & AssetSelection.assets("b")),
+        ("a or b", AssetSelection.assets("a") | AssetSelection.assets("b")),
+        ("+a", AssetSelection.assets("a").upstream(1)),
+        ("++a", AssetSelection.assets("a").upstream(2)),
+        ("a+", AssetSelection.assets("a").downstream(1)),
+        ("a++", AssetSelection.assets("a").downstream(2)),
+        (
+            "a* and *b",
+            AssetSelection.assets("a").downstream() and AssetSelection.assets("b").upstream(),
+        ),
+        ("sinks(a)", AssetSelection.assets("a").sinks()),
+        ("roots(c)", AssetSelection.assets("c").roots()),
+        ("tag:foo", AssetSelection.tag("foo", "")),
+        ("tag:foo=bar", AssetSelection.tag("foo", "bar")),
+        ('owner:"owner@owner.com"', AssetSelection.owner("owner@owner.com")),
+        ("group:my_group", AssetSelection.groups("my_group")),
+        ("kind:my_kind", AssetSelection.tag(f"{KIND_PREFIX}my_kind", "")),
+    ],
+)
+def test_antlr_visit_basic(selection_str, expected_assets):
+    # a -> b -> c
+    @asset(tags={"foo": "bar"}, owners=["team:billing"])
+    def a(): ...
+
+    @asset(deps=[a], kinds={"python", "snowflake"})
+    def b(): ...
+
+    @asset(
+        deps=[b],
+        group_name="my_group",
+    )
+    def c(): ...
+
+    defs = [a, b, c]
+
+    assert AntlrAssetSelectionParser(selection_str).asset_selection.resolve(
+        defs
+    ) == expected_assets.resolve(defs)
+
+
+def test_code_location() -> None:
+    @asset
+    def my_asset(): ...
+
+    # Selection can be instantiated.
+    selection = AntlrAssetSelectionParser("code_location:code_location1").asset_selection
+
+    assert selection == CodeLocationAssetSelection(selected_code_location="code_location1")
+
+    # But not resolved.
+    with pytest.raises(NotImplementedError):
+        selection.resolve([my_asset])
